build: resolve 2 instances of -Wimplicit-int-conversion
clang 19 on FreeBSD 14.3 warns:

```
IMAPMessage.cpp:112:11: warning: higher order bits are zeroes after implicit conversion [-Wimplicit-int-conversion]
  112 |           m_size(-1U),
      |                 ~^~~
IMAPMessage.cpp:183:16: warning: higher order bits are zeroes after implicit conversion [-Wimplicit-int-conversion]
  183 |         if (m_size == -1U) {
      |                    ~~ ^~~
```

-1U is UINT_MAX; assigning that to size_t does not change the value
(stays UINT_MAX, does not become SIZE_MAX).

What is really intended here is ``size_t(-1)`` (or syntactically
equivalents). This is equal to ``size_t(0) - 1`` as per
https://en.cppreference.com/w/c/language/conversion.html § Integer
conversions ("repeatedly subtracted"), and because unsigned integers
implement modulo arithmetic, we get SIZE_MAX as desired.

+++ src/vmime/net/imap/IMAPMessage.cpp
@@ -109,7 +109,7 @@ IMAPMessage::IMAPMessage(
 )
 	: m_folder(folder),
 	  m_num(num),
-	  m_size(-1U),
+	  m_size(-1),
 	  m_flags(FLAG_UNDEFINED),
 	  m_expunged(false),
 	  m_modseq(0),
@@ -180,7 +180,7 @@ vmime_uint64 IMAPMessage::getModSequence() const {
 
 size_t IMAPMessage::getSize() const {
 
-	if (m_size == -1U) {
+	if (m_size == static_cast<size_t>(-1)) {
 		throw exceptions::unfetched_object();
 	}
